+++ packages/bsky/src/api/index.ts
@@ -49,10 +49,10 @@ import registerPush from './app/bsky/notification/registerPush'
 import updateSeen from './app/bsky/notification/updateSeen'
 import getConfig from './app/bsky/unspecced/getConfig'
 import getPopularFeedGenerators from './app/bsky/unspecced/getPopularFeedGenerators'
+import getSuggestedStarterPacks from './app/bsky/unspecced/getSuggestedStarterPacks'
 import getTaggedSuggestions from './app/bsky/unspecced/getTaggedSuggestions'
 import getTrendingTopics from './app/bsky/unspecced/getTrendingTopics'
 import getTrends from './app/bsky/unspecced/getTrends'
-import getSuggestedStarterPacks from './app/bsky/unspecced/getSuggestedStarterPacks'
 import getUnspeccedSuggestedFeeds from './app/bsky/unspecced/getSuggestedFeeds'
 import getAccountInfos from './com/atproto/admin/getAccountInfos'
 import getSubjectStatus from './com/atproto/admin/getSubjectStatus'
@@ -112,6 +112,8 @@ export default function (server: Server, ctx: AppContext) {
   getTrends(server, ctx)
   getSuggestedStarterPacks(server, ctx)
   getUnspeccedSuggestedFeeds(server, ctx)
+  getTrends(server, ctx)
+  getSuggestedStarterPacks(server, ctx)
   getLabelerServices(server, ctx)
   searchActors(server, ctx)
   searchActorsTypeahead(server, ctx)