+++ packages/bsky/src/lexicon/types/app/bsky/unspecced/getSuggestedStarterPacks.ts
@@ -18,15 +18,13 @@ const is$typed = _is$typed,
 const id = 'app.bsky.unspecced.getSuggestedStarterPacks'
 
 export interface QueryParams {
-  /** DID of the account making the request (not included for public/unauthenticated queries). */
-  viewer?: string
   limit: number
 }
 
 export type InputSchema = undefined
 
 export interface OutputSchema {
-  starterPacks: AppBskyGraphDefs.StarterPackViewBasic[]
+  starterPacks: AppBskyGraphDefs.StarterPackView[]
 }
 
 export type HandlerInput = undefined