+++ packages/dev-env/CHANGELOG.md
@@ -1,5 +1,87 @@
 # @atproto/dev-env
 
+## 0.3.112
+
+### Patch Changes
+
+- Updated dependencies [[`04b6230cd`](https://github.com/bluesky-social/atproto/commit/04b6230cd2fbfe4a06cb00ab8ccb8e6c87c6c546)]:
+  - @atproto/bsky@0.0.135
+  - @atproto/api@0.14.18
+  - @atproto/pds@0.4.119
+  - @atproto/ozone@0.1.97
+
+## 0.3.111
+
+### Patch Changes
+
+- Updated dependencies [[`bdbd3c3e3`](https://github.com/bluesky-social/atproto/commit/bdbd3c3e3f8fe8476a3fecac73810554846c938f), [`2b7efb6cb`](https://github.com/bluesky-social/atproto/commit/2b7efb6cb1c93a108570efdafe9d9ec3f1018dfa), [`b0a0f1484`](https://github.com/bluesky-social/atproto/commit/b0a0f1484378adeb5e2aa20b9b6ff2c2eca0f740), [`0eea698be`](https://github.com/bluesky-social/atproto/commit/0eea698bef76520ae4cc0e1f2efbb588a0459556), [`bdbd3c3e3`](https://github.com/bluesky-social/atproto/commit/bdbd3c3e3f8fe8476a3fecac73810554846c938f)]:
+  - @atproto/api@0.14.17
+  - @atproto/bsky@0.0.134
+  - @atproto/pds@0.4.118
+  - @atproto/ozone@0.1.96
+  - @atproto/sync@0.1.19
+  - @atproto/bsync@0.0.17
+  - @atproto/crypto@0.4.4
+  - @atproto/xrpc-server@0.7.14
+
+## 0.3.110
+
+### Patch Changes
+
+- Updated dependencies [[`652894308`](https://github.com/bluesky-social/atproto/commit/65289430806976ec13177ed9c9f0e883e8f9330c)]:
+  - @atproto/ozone@0.1.95
+  - @atproto/bsky@0.0.133
+  - @atproto/api@0.14.16
+  - @atproto/pds@0.4.117
+
+## 0.3.109
+
+### Patch Changes
+
+- Updated dependencies [[`b4ab5011b`](https://github.com/bluesky-social/atproto/commit/b4ab5011bcc64f9f05122a8773806af8e0c13146)]:
+  - @atproto/api@0.14.15
+  - @atproto/pds@0.4.116
+  - @atproto/bsky@0.0.132
+  - @atproto/ozone@0.1.94
+
+## 0.3.108
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atproto/pds@0.4.115
+  - @atproto/bsky@0.0.131
+  - @atproto/ozone@0.1.93
+
+## 0.3.107
+
+### Patch Changes
+
+- Updated dependencies [[`9a05892f6`](https://github.com/bluesky-social/atproto/commit/9a05892f6fd405bf6bb96c9c8d2a9a89d5e94bc5)]:
+  - @atproto/ozone@0.1.93
+  - @atproto/bsky@0.0.131
+  - @atproto/api@0.14.14
+  - @atproto/pds@0.4.114
+
+## 0.3.106
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atproto/pds@0.4.113
+  - @atproto/bsky@0.0.130
+  - @atproto/ozone@0.1.92
+
+## 0.3.105
+
+### Patch Changes
+
+- Updated dependencies [[`076c2f987`](https://github.com/bluesky-social/atproto/commit/076c2f9872387217806624306e3af08878d1adcd)]:
+  - @atproto/ozone@0.1.92
+  - @atproto/api@0.14.13
+  - @atproto/pds@0.4.112
+  - @atproto/bsky@0.0.130
+
 ## 0.3.104
 
 ### Patch Changes